+++ example/basic_logger_example.dart
@@ -8,14 +8,19 @@ void main() {
   final basicLogger = BasicLogger('main');
 
   // attach developer log
-  /// basicLogger.attachLogger(DevOutputLogger(parentName: basicLogger.name));
+  basicLogger.attachLogger(DevOutputLogger(parentName: basicLogger.name));
 
-  // attach console log
-  final consoleLogger = basicLogger
-      .attachLogger(ConsoleOutputLogger(parentName: basicLogger.name));
+  // attach output log, alias console
+  // selfonly, when true, filter by selfname. Otherwise, output all.
+  final consoleLogger = basicLogger.attachLogger(OutputLogger(
+    basicLogger.name,
+    selfname: 'console',
+    // selfonly: true,
+  ));
+
   consoleLogger.info('console output 1234');
-
   basicLogger.debug('debug a1'); // output to all attach instance
+
   basicLogger.info('info a1'); // output to all attach instance
 
   // output buffer to all attach instance, not include detach instance